refactor: ssu_catch_plugin 모듈의 이름을 ssu_catch로 변경
